requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# helium-explorer-scraper/framework.py: 3
python_dotenv == 0.19.2

# helium-explorer-scraper/scraper.py: 3
requests == 2.25.1

# helium-explorer-scraper/scraper.py: 4
xlwt == 1.3.0
